startup_agile/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
pytest
psycopg2-binary
sqlalchemy
